requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==2.2.6
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-yasg==1.17.0
idna==2.8
importlib-metadata==1.3.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jwt==0.6.1
kafka-python==1.4.7
kombu==4.6.7
MarkupSafe==1.1.1
more-itertools==8.0.2
openapi-codec==1.3.2
packaging==19.2
pipreqs==0.4.9
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.2
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
xlrd==1.2.0
yarg==0.1.9
zipp==0.6.0
redis==3.4.1
pandas==1.0.3
